==> scripts/01_get_data.ipynb <==
<br>

# Introdução


#!pip3 install traquitanas --upgrade


import os
import time
import shutil
from datetime import date


# from traquitanas import net


from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC


from paths import *
from driver import *


<br>

# Funções


<br>

## Cria Driver


# download_path = os.path.abspath(os.path.join(os.getcwd(), bruto_path, 'temp'))
# print(download_path)

# driver = net.scraping.create_driver(
#     download_path,
#     headless=False,
#     adds_path=os.path.join('..', 'scrapy', 'adds'),
#     log_path=os.path.join('..', 'scrapy', 'logs')
# )


<br>

## _Download_ dos dados


# Vai pra site "Aplicações"
url = "http://www.aplicacoes.daee.sp.gov.br/usosrec/fchweb.html"
driver.get(url)


# Vai pra subsite "Baixar Usos p/ Bacia Hidrográfica"
url = 'http://www.aplicacoes.daee.sp.gov.br/usosrec/Daeewebbacia1.asp'
driver.get(url)


# Seleciona Bacia
ugrhi = Select(driver.find_element(by=By.XPATH, value="//*[@name='bacia1']"))
ugrhi.select_by_visible_text(
    '99 - TODAS AS BACIAS HIDROGRAFICAS DO ESTADO DE SAO PAULO'
)


# Clica em Selecionar (Esperar algum tepo)
driver.find_element(by=By.XPATH, value="//*[@value='SELECIONAR']").click()


# Espera algum componente aparecer
time.sleep(5)


# Fecha janela inicial
driver.switch_to.window(driver.window_handles[0])
driver.close()

# Muda para janela principal
driver.switch_to.window(driver.window_handles[0])


WebDriverWait(driver, 80).until(
    EC.visibility_of_element_located((By.XPATH, '//*[contains(@class, "caixa1")]'))
)


# Scrool Down
driver.execute_script('window.scrollTo(0, document.body.scrollHeight);')


# Clicar para Download
driver.find_element(by=By.XPATH, value='//*[contains(@class, "caixa1")]').click()


download_ok = False
n_tries = 0

while not download_ok and n_tries < 10:
    list_files = os.listdir(os.path.join(bruto_path, 'temp'))

    if len(list_files) > 1:
        time.sleep(1)

    if len(list_files) == 1:
        hoje = date.today().strftime("%Y.%m.%d")
        shutil.move(
            os.path.join(bruto_path, 'temp', list_files[0]),
            os.path.join(bruto_path, f'outorgas_{hoje}.txt'),
        )

    if len(list_files) == 0:
        # Espera download terminar
        driver.quit()
        time.sleep(1)
        os.chmod(os.path.join(bruto_path, 'temp'), 0o777)
        shutil.rmtree(
            os.path.join(bruto_path, 'temp'),
            onerror=print('ee'),
            ignore_errors=True,
        )
        download_ok = True

    n_tries += 1
